Criando a tabela taks através do Knex e criando os primeiros tests com Jest
