+++ src/app.ts
@@ -1,0 +1,9 @@
+import express, { Request, Response } from 'express';
+
+const app = express()
+
+app.get('/', (req: Request, res: Response) => {
+    res.send('Rodando')
+})
+
+export default app
