+++ src/server.ts
@@ -1,9 +1,3 @@
-import express, { Request, Response } from 'express';
-
-const app = express()
-
-app.get('/', (req: Request, res: Response) => {
-    res.send('Rodando')
-})
+import app from './app'
 
 app.listen(8080)
